backend/requirements.txt
# Soundromeda backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.25
aiosqlite>=0.19.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# Pipeline: audio features + layout
librosa>=0.10.0
soundfile>=0.12.0
umap-learn>=0.5.0
numpy>=1.24.0
joblib>=1.3.0
